+++ src/ExportManager/RecentExport.cs
@@ -35,7 +35,7 @@ namespace SCaddins.ExportManager
                 {
                     foreach (var vsi in viewSetItems)
                     {
-#if REVIT2024
+#if REVIT2024 || REVIT2025
                         doc.Delete(new ElementId((long)vsi.Id));
 #else
                         doc.Delete(new ElementId(vsi.Id));
@@ -61,7 +61,7 @@ namespace SCaddins.ExportManager
                 {
                     if (transaction.Start("Delete oldest user export set") == TransactionStatus.Started)
                     {
-#if REVIT2024
+#if REVIT2024 || REVIT2025
                         doc.Delete(new Autodesk.Revit.DB.ElementId((long)oldest.Id));
 #else
                         doc.Delete(new Autodesk.Revit.DB.ElementId(oldest.Id));