+++ src/ExportManager/SegmentedSheetName.cs
@@ -33,7 +33,7 @@ namespace SCaddins.ExportManager
             Hooks = new Collection<string>();
             // ReSharper disable once StringLiteralTypo
             Name = "YYYYMMDD-AD-NNN[R]";
-#if REVIT2022 || REVIT2023 || REVIT2024
+#if REVIT2022 || REVIT2023 || REVIT2024 || REVIT2025
             PDFExportOptions = null;
 #endif
         }
@@ -68,7 +68,7 @@ namespace SCaddins.ExportManager
             }
         }
 
-#if REVIT2022 || REVIT2023 || REVIT2024
+#if REVIT2022 || REVIT2023 || REVIT2024 || REVIT2025
         public PDFExportOptions PDFExportOptions { get; set; }
 #endif