+++ src/ExportManager/Sheet.cs
@@ -205,7 +205,7 @@ namespace SCaddins.ExportManager
         {
             get
             {
-#if REVIT2022 || REVIT2023 || REVIT2024
+#if REVIT2022 || REVIT2023 || REVIT2024 || REVIT2025
                 return SegmentedFileName.Name;
 #else
                 return printSetting != null ? printSetting.Name : string.Empty;
@@ -307,7 +307,7 @@ namespace SCaddins.ExportManager
 
         public string SheetRevision
         {
-#if REVIT2022 || REVIT2023 || REVIT2024
+#if REVIT2022 || REVIT2023 || REVIT2024 || REVIT2025
             get
             {
                 return sheetRevision != string.Empty ? sheetRevision : "-";
@@ -594,7 +594,7 @@ namespace SCaddins.ExportManager
 
         public override string ToString()
         {
-#if REVIT2022 || REVIT2023 || REVIT2024
+#if REVIT2022 || REVIT2023 || REVIT2024 || REVIT2025
             string printSetting = SegmentedFileName.Name;
 #endif
 
@@ -725,7 +725,7 @@ namespace SCaddins.ExportManager
             }
             appearsInSheetList = this.Sheet.get_Parameter(BuiltInParameter.SHEET_SCHEDULED).AsInteger() == 1;
             pageSize = PrintSettings.GetSheetSizeAsString(this);
-#if !REVIT2022 && !REVIT2023 && !REVIT2024
+#if !REVIT2022 && !REVIT2023 && !REVIT2024 || REVIT2025
             printSetting = PrintSettings.GetPrintSettingByName(doc, pageSize, forceRasterPrint);
             if (printSetting == null)
             {
@@ -733,7 +733,7 @@ namespace SCaddins.ExportManager
             }
 #endif
             verified = true;
-#if !REVIT2022 && !REVIT2023 && !REVIT2024
+#if !REVIT2022 && !REVIT2023 && !REVIT2024 && !REVIT2025
             ValidPrintSettingIsAssigned = printSetting != null;
 #else
             ValidPrintSettingIsAssigned = true;
@@ -797,7 +797,7 @@ namespace SCaddins.ExportManager
 
         private string PopulateSegmentedFileName()
         {
-#if REVIT2022 || REVIT2023 || REVIT2024
+#if REVIT2022 || REVIT2023 || REVIT2024 || REVIT2025
             var opts = this.SegmentedFileName.PDFExportOptions;
             if (opts != null)
             {
@@ -809,7 +809,7 @@ namespace SCaddins.ExportManager
 
         private void SetExportName()
         {
-#if REVIT2022 || REVIT2023 || REVIT2024
+#if REVIT2022 || REVIT2023 || REVIT2024 || REVIT2025
             sheetRevision = sheet.get_Parameter(
                     BuiltInParameter.SHEET_CURRENT_REVISION).AsString();
 #else