+++ src/GridManager/Views/GridManagerView.xaml.cs
@@ -1,0 +1,10 @@
+namespace SCaddins.GridManager.Views
+{
+    public partial class GridManagerView
+    {
+        public GridManagerView()
+        {
+            InitializeComponent();
+        }
+    }
+}