+++ src/ModelSetupWizard/ModelSetupWizardSettings.Designer.cs
@@ -12,7 +12,7 @@ namespace SCaddins.ModelSetupWizard {
     
     
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
-    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("Microsoft.VisualStudio.Editors.SettingsDesigner.SettingsSingleFileGenerator", "16.3.0.0")]
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("Microsoft.VisualStudio.Editors.SettingsDesigner.SettingsSingleFileGenerator", "17.9.0.0")]
     internal sealed partial class ModelSetupWizardSettings : global::System.Configuration.ApplicationSettingsBase {
         
         private static ModelSetupWizardSettings defaultInstance = ((ModelSetupWizardSettings)(global::System.Configuration.ApplicationSettingsBase.Synchronized(new ModelSetupWizardSettings())));