+++ src/SpellChecker/SpellCheckerSettings.Designer.cs
@@ -12,7 +12,7 @@ namespace SCaddins.SpellChecker {
     
     
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
-    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("Microsoft.VisualStudio.Editors.SettingsDesigner.SettingsSingleFileGenerator", "16.3.0.0")]
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("Microsoft.VisualStudio.Editors.SettingsDesigner.SettingsSingleFileGenerator", "17.9.0.0")]
     internal sealed partial class SpellCheckerSettings : global::System.Configuration.ApplicationSettingsBase {
         
         private static SpellCheckerSettings defaultInstance = ((SpellCheckerSettings)(global::System.Configuration.ApplicationSettingsBase.Synchronized(new SpellCheckerSettings())));